+++ tests/testthat/test-comment_binary_operator.R
@@ -4,16 +4,19 @@ test_that("%:)% adds comments",{
 
     minimal <- construct("a", "b", "c" %:)% "A love note to your future self")
     
-    x1 <- structure("abc", class = c("regexr", "character"), regex = list(
-            "a", "b", structure("c", class = c("regex", "character"), comment = "A love note to your future self")), comments = list(
+    x1 <- structure("abc", class = c("regexr", "character"), subs = list(
+            "a", "b", structure("c", class = c("subcom", "character"), 
+                comment = "A love note to your future self")), comments = list(
             NULL, NULL, "A love note to your future self"))
     
     expect_equivalent(x1, minimal)
     
-    x2 <- structure(list(class = c("regexr", "character"), regex = list(
-            "a", "b", structure("c", class = c("regex", "character"), comment = "A love note to your future self")), 
-            comments = list(NULL, NULL, "A love note to your future self")), .Names = c("class", 
-        "regex", "comments"))
+    x2 <- structure(list(class = c("regexr", "character"), subs = list(
+            "a", "b", structure("c", class = c("subcom", "character"), 
+                comment = "A love note to your future self")), 
+            comments = list(NULL, NULL, "A love note to your future self")), 
+        .Names = c("class", 
+        "subs", "comments"))
     
     
     expect_equivalent(x2, attributes(minimal))