+++ tests/testthat/test-construct.R
@@ -18,8 +18,8 @@ test_that("construct makes a regular expression of class regexr with appropriate
     )
 
     m2 <- structure("\\s+(?<=(foo))(;|:)\\s*[a]s th[atey]", class = c("regexr", 
-        "character"), regex = structure(list(space = structure("\\s+", class = c("regex", 
-        "character"), comment = "I see"), simp = "(?<=(foo))", or = structure("(;|:)\\s*", class = c("regex", 
+        "character"), subs = structure(list(space = structure("\\s+", class = c("subcom", 
+        "character"), comment = "I see"), simp = "(?<=(foo))", or = structure("(;|:)\\s*", class = c("subcom", 
         "character"), comment = "comment on what this does"), "[a]s th[atey]"), .Names = c("space", 
         "simp", "or", "")), comments = structure(list(space = "I see", 
             simp = NULL, or = "comment on what this does", NULL), .Names = c("space",