+++ tests/testthat/test-summary.R
@@ -3,15 +3,16 @@ context("Checking summary")
 test_that("summary outputs pieces needed for printing",{
 
     minimal <- structure("abc", class = c("regexr", "character"), subs = list(
-        "a", "b", structure("c", class = c("regex", "character"), comment = "A love note to your future self")), comments = list(
+        "a", "b", structure("c", class = c("subcom", "character"), 
+            comment = "A love note to your future self")), comments = list(
         NULL, NULL, "A love note to your future self"))
 
    
     x <- structure(list(`NA` = structure(list(comment = NULL, subs = "a"), .Names = c("comment", 
-       "regex")), Foo = structure(list(comment = NULL, subs = "b"), .Names = c("comment", 
-        "regex")), Bar = structure(list(comment = "A love note to your future self", 
-        subs = structure("c", class = c("regex", "character"), comment = "A love note to your future self")), .Names = c("comment", 
-        "regex"))), .Names = c(NA, "Foo", "Bar"), class = "summary_regexr", subs = "abc")
+       "subcom")), Foo = structure(list(comment = NULL, subs = "b"), .Names = c("comment", 
+        "subcom")), Bar = structure(list(comment = "A love note to your future self", 
+        subs = structure("c", class = c("subcom", "character"), comment = "A love note to your future self")), .Names = c("comment", 
+        "subcom"))), .Names = c(NA, "Foo", "Bar"), class = "summary_regexr", subs = "abc")
 
     
     expect_equivalent(summary(minimal), x)